refactor: remove util func from ConfluxTestFrameworkForContract
