+++ tests/test_framework/contracts.py
@@ -14,32 +14,8 @@ class ConfluxTestFrameworkForContract(ConfluxTestFramework):
         self.num_nodes = 1
         self.conf_parameters["executive_trace"] = "true"
 
-    def assert_tx_exec_error(self, tx_hash, err_msg):
-        self.client.wait_for_receipt(tx_hash)
-        receipt = self.client.get_transaction_receipt(tx_hash)
-        assert_equal(
-            receipt["txExecErrorMsg"],
-            err_msg
-        )
-
     def before_test(self):
-        super().before_test()
         self.rpc = self.nodes[0].rpc
         self.enable_max_priority_fee_per_gas()
         self.setup_w3()
         self.w3 = self.cw3
-
-    def initialize_accounts(self, number = 10, value = 100):
-        """
-        Not recommended now. It is now recommended to use `self._add_genesis_account` 
-        during param setting phase to add genesis accounts.
-        
-        The generated accounts can be used from self.core_accounts or self.evm_accounts.
-        """
-        def initialize_new_account():
-            acct = self.cfx.account.create()
-            if value > 0:
-                self.cfx_transfer(acct.hex_address, value = value)
-            return acct
-        
-        return [initialize_new_account() for _ in range(number)]
